+++ src/main/java/FunctionLayer/LogicFacade.java
@@ -53,6 +53,6 @@ public interface LogicFacade {
     
     ArrayList<Order> getOrdersFromCustomer(int customerId) throws CarportException;
     
-    LineItem createLineItem(int materialId, int orderId, int qty, int length, int width, int height, String comment)throws CarportException;
+    LineItem createLineItem(int materialId, int orderId, int qty, double length, double width, double height, String comment)throws CarportException;
             
 }
