+++ src/main/java/FunctionLayer/LogicFacadeImplementation.java
@@ -121,8 +121,8 @@ public class LogicFacadeImplementation implements LogicFacade {
     }
 
     @Override
-    public LineItem createLineItem(int materialId, int orderId, int qty, int length,
-            int width, int height, String comment) throws CarportException {
+    public LineItem createLineItem(int materialId, int orderId, int qty, double length,
+            double width, double height, String comment) throws CarportException {
         LineItemMapper lim = new LineItemMapper();
         LineItem lineItem = new LineItem(materialId, orderId, qty, length, width, height, comment);
         lim.createLineItem(lineItem);
